moved objectDetection model to models/objectDetection
