+++ src/utils/types.ts
@@ -1,13 +1,12 @@
-import * as ssd from '@tensorflow-models/coco-ssd';
+export type Bbox = [number, number, number, number] | undefined;
+
+export interface ICoords {
+    x: number;
+    y: number;
+}
 
 export type DetectionImage =
     | HTMLVideoElement
     | HTMLImageElement
     | HTMLCanvasElement
     | ImageData;
-
-export type Bbox = [number, number, number, number] | undefined;
-
-export type ModelConfig = ssd.ObjectDetectionBaseModel;
-
-export type DetectionConfig = number;
